+++ crates/cast/Cargo.toml
@@ -53,8 +53,8 @@ alloy-signer.workspace = true
 alloy-sol-types.workspace = true
 alloy-transport.workspace = true
 alloy-zksync.workspace = true
-
 zksync_types.workspace = true
+zksync-tx-types.workspace = true
 
 chrono.workspace = true
 eyre.workspace = true
