+++ crates/cast/bin/cmd/call/zksync.rs
@@ -4,8 +4,8 @@ use alloy_rpc_types::TransactionRequest;
 use alloy_serde::WithOtherFields;
 use alloy_sol_types::SolCall;
 use alloy_zksync::network::transaction_request::TransactionRequest as ZkTransactionRequest;
-use cast::ZkTransactionOpts;
 use eyre::Result;
+use zksync_tx_types::ZkTransactionOpts;
 
 /// Converts the given tx request to be a full ZkSync transaction request with fee estimation
 pub async fn convert_tx(
