+++ crates/cli/Cargo.toml
@@ -31,6 +31,7 @@ alloy-provider.workspace = true
 alloy-rlp.workspace = true
 alloy-chains.workspace = true
 alloy-zksync.workspace = true
+zksync-tx-types.workspace = true
 
 clap = { version = "4", features = ["derive", "env", "unicode", "wrap_help"] }
 color-eyre.workspace = true
